Working but very slow implementation
