+++ flux/try.jl
@@ -1,47 +1,30 @@
 using Flux, Flux.Tracker
 using LinearAlgebra
 using Flux.Tracker: update!
+using Word2Vec
 
-#define parameters
-#10000 - 200 dimensional vecotrs
-ov = param(randn(10000,200)/3)
+sentences = tokenize_document("./data/shakespeare_all.txt")
 
-# define functions
+vocab_hash, vocab, pq, data = process_input(sentences, 3, 5, :cbow)
 
-# nodeprob: returns probability you are going the right way at a node
-# ov = the output vector matrix a Flux param
-# h = internal representation
-# node = the current node number
-# branch is the direction to branch from the node
-function nodeprob(ov, h, node, branch)
-    v = ov[node,:]
-    branch == 1 ? ◫ = 1 : ◫ = -1
-    return σ(◫*dot(v,h))
-end
+data = minibatch(data, 500)
 
-# prob: returns probability(output word|input word)
-# h = internal representation
-# nodes = array of nodes in path to output word from Huffman tree
-# branches = array of directions (0,1) to branch from each node in nodes
-function prob(ov, h, nodes, branches)
-    e = reduce(*, map((node, branch)->(nodeprob(ov, h, node, branch)), nodes, branches))
-end
+wv = WordVectors(25, vocab, vocab_hash)
 
-# loss is the -log of the probability(output word|input word)
-loss1(ov, h, nodes, branches) = -1*log(prob(ov, h, nodes, branches))
-loss2(ov, h, nodes, branches) = -1*reduce(+, map((node, branch)->log((nodeprob(ov, h, node, branch))), nodes, branches))
+hs = HeirarchicalSoftmax(25, pq, vocab_hash, vocab)
 
+loss(z) = -1*log(hs(wv(z...)))
 
-# try it out
-h = param(randn(200)/3)
-nodes = [200,300,400,500]
-branches = [0,0,1,0]
+ps = params(hs.tree_vectors, wv.vectors)
 
-for i in 1:100
-    # println(loss1(ov, h, nodes, branches))
-    grad1 = Tracker.gradient(()->loss1(ov, h, nodes, branches), Params([ov, h]))
-    a = grad1[ov]
-    b = grad1[h]
-    update!(ov, -0.1*a)
-    update!(h, -0.1*b)
+using ProgressMeter
+
+@showprogress 1 for i in data
+        l = loss(i)
+        #println(l)
+        grad1 = Tracker.gradient(()->l, ps)
+        a = grad1[hs.tree_vectors]
+        b = grad1[wv.vectors]
+        update!(hs.tree_vectors, -0.1*a)
+        update!(wv.vectors, -0.1*b)
 end
