+++ src/Word2Vec.jl
@@ -1,9 +1,6 @@
 module Word2Vec
 
 using Reexport
-# using IterTools
-# using DataStructures
-# using LinearAlgebra
 
 include("./doctools/tokenize.jl")
 export tokenize_document
@@ -19,30 +16,17 @@ include("contextdata/ContextDataTypes.jl")
 
 include("heirarchical_softmax.jl")
 export HeirarchicalSoftmax
-#
-# include("wordvectors.jl")
-# export
-#     # types
-#     WordVectors,
-#
-#     # functions
-#     wordvector
-#
-# include("network.jl")
-# export
-#     # types
-#     W2VNetwork,
-#
-#     # functions
-#     word2vec, train!, extract_vectors! # still need to be implemented
-#
-# include("vectorinteract.jl")
-#     export
-#     # functions
-#     index, size,
-#     vocabulary, in_vocabulary,
-#     get_vector, get_cluster, get_words,
-#     cosine, cosine_similar_words,
-#     analogy, analogy_words, similarity
+
+include("wordvectors.jl")
+export
+    # types
+    WordVectors,
+
+    # functions
+    wordvector
+    index, size,
+    vocabulary, in_vocabulary,
+    get_vector, cosine, cosine_similar_words,
+    analogy, analogy_words, similarity
 
 end
