+++ src/contextdata/wrappers.jl
@@ -3,6 +3,11 @@ struct DownsampledData{T<:TextData, F<:AbstractFloat} <: TextData
     threshold::F
 end
 
+function downsample(data::TextData, threshold::AbstractFloat)
+    DownsampledData(data, threshold)
+end
+
+
 function Base.iterate(data::DownsampledData)
     error("you didn't implement this yet, dummy")
 end
@@ -11,25 +16,24 @@ struct MiniBatchData{T<:Iterators.PartitionIterator{<:TextData}}
     minidata::T
 end
 
-function MiniBatchData(data::TextData, n::Integer)
+function minibatch(data::TextData, n::Integer)
     minidata = (Iterators.partition(data, n))
-    return MiniBatchData(minidata)
+    MiniBatchData(minidata)
 end
 
 Base.length(data::MiniBatchData) = length(data.minidata)
 
 function Base.iterate(data::MiniBatchData, state = 1)
-    if state < 1 || state > length(data) + 1
-        error("Invalid iterator state parameter")
-    elseif state == length(data) + 1
+    temp = iterate(data.minidata, state)
+    if temp == nothing
         return nothing
     else
-        (temp, newstate) = iterate(data.minidata, state)
+        (temp, newstate) = temp
         temp = Iterators.Stateful(temp)
         X, Y = iterate(temp)[1]
         for (i, j) in temp
             X = hcat(X, i)
-            Y = hcat(Y, j)
+            Y = vcat(Y, j)
         end
     end
     return ((X, Y), newstate)
